requirements.txt
absl-py==1.0.0
appnope==0.1.2
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bleach==3.3.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.14.5
charset-normalizer==2.0.12
click==8.1.2
cycler==0.10.0
Cython==0.29.23
debugpy==1.5.1
decorator==5.0.9
defusedxml==0.7.1
docker-pycreds==0.4.0
entrypoints==0.3
et-xmlfile==1.1.0
flatbuffers==2.0
gast==0.5.3
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.6.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h5py==3.6.0
idna==3.3
imageio==2.16.0
importlib-metadata==4.11.3
ipykernel==6.4.2
ipython==7.23.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
joblib==1.1.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
libclang==13.0.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.0.3
matplotlib-inline==0.1.2
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.20.3
oauthlib==3.2.0
opencv-python==4.5.5.62
openpyxl==3.0.9
opt-einsum==3.3.0
packaging==20.9
pandas==1.3.4
pandocfilters==1.4.3
parso==0.8.2
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
prometheus-client==0.10.1
promise==2.3
prompt-toolkit==3.0.18
protobuf==3.20.0
psutil==5.9.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.3
PyYAML==6.0
pyzmq==22.0.3
qtconsole==5.1.0
QtPy==1.9.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0
scipy==1.7.0
seaborn==0.11.2
Send2Trash==1.5.0
sentry-sdk==1.5.9
setproctitle==1.2.2
shortuuid==1.0.8
six==1.16.0
sklearn==0.0
smmap==5.0.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
terminado==0.10.0
testpath==0.5.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.0.0
torch==1.11.0
tornado==6.1
traitlets==5.0.5
typing_extensions==4.1.1
urllib3==1.26.9
wandb==0.12.14
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.1
widgetsnbextension==3.5.1
wrapt==1.14.0
zipp==3.8.0
